requirements.txt
annotated-types==0.7.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
deepspeed==0.17.6
dnspython==2.8.0
einops==0.8.1
eventlet==0.40.3
filelock==3.19.1
flask==3.1.2
flask-socketio==5.5.1
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
hjson==3.1.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
ninja==1.13.0
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pillow==11.3.0
psutil==7.1.0
py-cpuinfo==9.0.0
pydantic==2.11.9
pydantic-core==2.33.2
python-engineio==4.12.2
python-socketio==5.13.0
requests==2.32.5
setuptools==80.9.0
simple-websocket==1.1.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
websocket-client==1.8.0
werkzeug==3.1.3
wsproto==1.2.0
